Current

+++ images/03.ServiceImages/certs/home/engines/scripts/engine/create_csr.sh
@@ -1,41 +1,14 @@
 #!/bin/sh
 sudo -n /home/engines/scripts/engine/sudo/_fix_perms.sh
 . /home/engines/scripts/engine/cert_dirs.sh
-isUserCert=0
+
 cert_name=`echo $common_name | sed "s/$.//"` 
 export cert_name
-
-#if test -z $cert_type 
-# then
-#  cert_type=generated
-#fi
-#
-#if test $cert_type = user
-# then
-#    cert_type=generated
-#	StorePref=user
-#	isUserCert=1
-#  elif test $cert_type = external_ca
-#  then
-#   StorePref=/
-#  else
-#   StorePref=${container_type}s/${parent_engine}
-#fi
  
 if test -z $hostname
  then
   hostname=$common_name
 fi
-
-if test -z $key_dir
- then
-    key_dir=$StoreRoot/$cert_type/keys/${StorePref}
-fi
-if test -z $cert_dir
- then
-cert_dir=$StoreRoot/$cert_type/certs/${StorePref}
-fi
-
 
 if ! test -d $key_dir
  then
@@ -46,7 +19,15 @@ then
  mkdir -p $cert_dir
 fi
   
-setup_dir=/home/certs/saved
+if ! test -d $setup_dir
+ then
+  mkdir -p $setup_dir
+fi
+
+if ! test -d $pending_csr_dir 
+ then
+  mkdir -p $pending_csr_dir
+fi
 
  
 if test -z $wild
@@ -56,7 +37,7 @@ fi
 
 if test -z "$country" -a -z "$state" -a -z "$organisation"
  then
- . $StoreRoot/default_cert_details
+  load_cert_defaults
 fi
 
 echo $country >$setup_dir/${cert_name}_setup
@@ -107,15 +88,25 @@ if ! test -z "$alt_names"
  	done
 fi
 
+resolve_cert_dir
+resolve_key_dir
+
 cat /etc/ssl/openssl.cnf $setup_dir/${cert_name}_config >$setup_dir/${cert_name}_config_full
 openssl genrsa -out  $key_dir/${cert_name}.key.tmp 2048
 
+if ! test -f  $key_dir/${cert_name}.key.tmp
+ then
+  echo '{"status":"error","message":"Failed to create key ' $key_dir/${cert_name}.key'" }'
+  exit 3
+fi  
+
+
 openssl req -new  -key $key_dir/${cert_name}.key.tmp -out $pending_csr_dir/${cert_name}.csr -config $setup_dir/${cert_name}_config
-cp $setup_dir/${cert_name}_config $cert_dir
+
 if ! test -f $pending_csr_dir/${cert_name}.csr 
   then
- 	echo "Failed to Create CSR"
- 	rm $key_dir/${common_name}.key.tmp
- 	exit 127
+ 	echo '{"status":"error","message":"Failed to Create CSR '$pending_csr_dir/${cert_name}.csr'"}'
+ 	rm  $key_dir/${cert_name}.key.tmp
+ 	exit 3
 fi
  mv $key_dir/${common_name}.key.tmp $key_dir/${common_name}.key
